Update to solution to match changes to instructions

+++ WishList/Controllers/ItemController.cs
@@ -47,6 +47,8 @@ namespace WishList.Controllers
         {
             var user = _userManager.GetUserAsync(HttpContext.User).Result;
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
+            if (item == null)
+                return NotFound();
             if (item.User.Id != user.Id)
                 return Unauthorized();
             _context.Items.Remove(item);
